- update `readme.md` - rename folder to 'category_search'
