+++ docs/virtual-env.md
@@ -125,12 +125,12 @@ python3 ./manage.py test
 
 To run a single test, specify the app name, file structure to access the test, test class and test method name:
 ```
-python3 ./manage.py test <app_name>.<test_folder_name>.<test_file_name>.<test_class>.<test_method>
+python3 ./manage.py test tests.<test_file_name>.<test_class>.<test_method>
 ```
 
 For example:
 ```
-python3 ./manage.py test adviser.tests.test_search_view_function.ResultsPageWithBothOrgAndPostcodeTest.test_search_parameters_box_is_visible
+python3 ./manage.py test tests.test_fala_views_and_functions.ResultsPageWithBothOrgAndPostcodeTest.test_search_parameters_box_is_visible
 ```
 
 You can also omit the test method from the command and run the whole class.
